Added efi verification
